completed badge integration and made minor modifications
