--- badge.md
@@ -1,6 +1,0 @@
-
-## Table of Contents
-- [License](#License)
-
-## License
-![GNU GPLv2](https://img.shields.io/badge/License-GPLv2-blue.svg)
